Implement top up flow with invoices
